# swagger/docs/swagger.yaml
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Demo
  version: "1.0"
paths:
  /api/v1/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 登录
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 手机号
        in: formData
        name: mobile
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 注册
      tags:
      - auth
  /api/v1/user:
    get:
      consumes:
      - application/json
      description: 用户列表
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 用户列表
      tags:
      - user
  /api/v1/user/{id}:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 删除用户
      tags:
      - user
    get:
      consumes:
      - application/json
      description: 用户详情
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 用户详情
      tags:
      - user
    put:
      consumes:
      - application/json
      description: 修改用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 手机号
        in: query
        name: mobile
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 修改用户
      tags:
      - user
  /api/v1/user/add:
    post:
      consumes:
      - application/json
      description: 新增用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 手机号
        in: formData
        name: mobile
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: 新增用户
      tags:
      - user
swagger: "2.0"